Remove zones, fault trapping, and process set functions; rewrite env handling.
This is really several commits in one, because the parts all rely on each other,
and disentangling them didn't seem worthwhile.

The heart of it is an attempt to move towards removal of the Solaris-specific
header <sys/procfs.h>. This is not yet accomplished, but partially: most places
that included it no longer do, and a number of helper functions (notably
prset*()) to operate on Solaris libproc are gone. Zones are gone. Fault trapping
is gone (PTRACE_TRACEME does much the same job).

The other fix entangled with this is a near-rewrite of Penv_iter(), which no
longer works by digging the environment up using the old libproc functions and
iterating through it piece by piece, but read()s it out of /proc/$pid/environ in
one go and iterates through the result. It's a dramatic simplification.

Finally, a tiny fix to our dt_ioctl() declaration: Solaris ioctl() 'val' is
'int', Linux is 'unsigned long', and the new ioctls were overflowing 'int'.

+++ libdtrace/dt_impl.h
@@ -551,7 +551,7 @@ extern int dt_set_errno(dtrace_hdl_t *, int);
 extern void dt_set_errmsg(dtrace_hdl_t *, const char *, const char *,
     const char *, int, const char *, va_list);
 
-extern int dt_ioctl(dtrace_hdl_t *, int, void *);
+extern int dt_ioctl(dtrace_hdl_t *, unsigned long int, void *);
 extern int dt_status(dtrace_hdl_t *, processorid_t);
 extern long dt_sysconf(dtrace_hdl_t *, int);
 extern ssize_t dt_write(dtrace_hdl_t *, int, const void *, size_t);
